+++ main.go
@@ -3,7 +3,5 @@ package main
 import "raj/tasket/cmd"
 
 func main() {
-	cmd.RootCmd.AddCommand(cmd.AddCmd)
-
-	cmd.RootCmd.Execute()
+	cmd.Execute()
 }